requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.1

